requirements_current.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
backports.asyncio.runner==1.2.0
beautifulsoup4==4.12.2
bidict==0.23.1
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
crcmod==1.7
curl_cffi==0.11.4
cycler==0.12.1
Cython==0.29.37
debugpy==1.8.1
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
editdistance==0.8.1
email_validator==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fake-useragent==1.4.0
fastapi==0.111.0
fastapi-cli==0.0.4
feedparser==6.0.10
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.53.0
frozendict==2.4.6
frozenlist==1.4.1
fsspec==2023.12.0
future==1.0.0
gotrue==2.12.2
gradio==3.38.0
gradio_client==0.8.1
greenlet==3.1.1
grpcio==1.64.1
h11==0.14.0
h2==4.2.0
hdbscan==0.8.36
hpack==4.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.19.4
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.1.0
ipykernel==6.29.4
ipython==8.18.1
itsdangerous==2.2.0
jaconv==0.3.4
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
jiter==0.10.0
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kaldiio==2.18.0
kiwisolver==1.4.5
librosa==0.9.2
lightning-utilities==0.11.2
linkify-it-py==2.0.3
llvmlite==0.39.1
lxml==6.0.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numba==0.56.4
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.18.0
openai==1.92.2
openpyxl==3.1.5
orjson==3.10.5
oss2==2.18.6
outcome==1.3.0.post0
packaging==24.1
pandas==2.1.3
parso==0.8.4
peewee==3.17.7
pillow==10.3.0
pip==25.1.1
pip-check==2.8.1
platformdirs==4.2.2
playwright==1.48.0
pluggy==1.6.0
pooch==1.8.2
postgrest==1.1.1
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
py==1.11.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.1.2
pyreadline3==3.4.1
PySocks==1.7.1
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pytorch-lightning==2.3.0
pytorch-wpe==0.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
realtime==2.5.3
referencing==0.35.1
regex==2025.7.34
requests==2.31.0
resampy==0.4.3
retry==0.9.2
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.1
schedule==1.2.2
scikit-learn==1.5.0
scipy==1.13.1
selenium==4.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==49.2.1
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
sympy==1.12.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
terminaltables==3.1.10
threadpoolctl==3.5.0
tokenizers==0.15.0
tomli==2.2.1
toolz==0.12.1
torch-complex==0.4.3
torchmetrics==1.4.0.post0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.35.2
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.3
typing_extensions==4.14.0
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
umap==0.1.1
undetected-chromedriver==3.5.5
urllib3==2.2.1
uv==0.7.9
uvicorn==0.30.1
wcwidth==0.2.13
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.4
yfinance==0.2.63
zipp==3.19.2
